+++ drivers/media/tuners/si2157.c
@@ -680,8 +680,8 @@ static int si2157_get_rf_strength(struct dvb_frontend *fe,
 	}
 
 	c->strength.stat[0].scale = FE_SCALE_DECIBEL;
-	c->strength.stat[0].svalue = (s8) cmd.args[3] * 1000;
-	*signal_strength = (s8) cmd.args[3];
+	c->strength.stat[0].svalue = (s8)cmd.args[3] * 1000;
+	*signal_strength = cmd.args[3];
 
 	return 0;
 }
